requirements.txt
anyio==4.7.0
click==8.1.7
graphql-core==3.2.5
grpcio==1.68.1
grpcio-tools==1.68.1
h11==0.14.0
idna==3.10
protobuf==5.29.1
python-dateutil==2.9.0.post0
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
starlette==0.42.0
strawberry-graphql==0.254.0
typing_extensions==4.12.2
uvicorn==0.34.0
